# src/main/resources/static/openapi/v1.yaml
openapi: 3.1.0
info:
  title: rbgs.io API
  version: 1.0.0
  description: Foundation, Battle.net sessions, character profiles and season launch notifications. Other product endpoints are added with their owning features.
servers:
  - url: /api/v1
paths:
  /season-notifications:
    get:
      operationId: getSeasonNotificationStatus
      summary: Get launch state and establish an anonymous CSRF cookie
      responses:
        '200':
          description: Notification state, without provider credentials
          content:
            application/json:
              schema:
                type: object
                required: [seasonLive, subscriptionsAvailable]
                properties:
                  seasonLive: { type: boolean }
                  subscriptionsAvailable: { type: boolean }
    post:
      operationId: subscribeToSeasonLaunch
      summary: Request one launch notification, without a Battle.net account
      description: Requires X-XSRF-TOKEN from the XSRF-TOKEN cookie. The website field is a honeypot and must be empty. Repeated and already-confirmed addresses receive the same generic response. Confirmation links expire after 48 hours; re-requests are limited to one per address per 15 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, language, consent]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                language: { type: string, enum: [en, es] }
                consent: { type: boolean, const: true }
                website: { type: string, maxLength: 200, default: '' }
      responses:
        '202': { description: Request accepted; confirmation may be queued if required }
        '400': { description: Invalid email, locale or consent }
        '403': { description: Missing or invalid CSRF token }
        '409': { description: Season already started }
        '429': { description: Too many requests }
        '503': { description: Email delivery has not been configured }
  /season-notifications/confirm:
    post:
      operationId: confirmSeasonNotification
      summary: Confirm an email using a bearer token and a deliberate user action
      description: Requires X-XSRF-TOKEN. Does not require Battle.net. An already-confirmed active token succeeds idempotently.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SeasonNotificationLink' }
      responses:
        '204': { description: Email confirmed }
        '400': { description: Invalid, expired or cancelled confirmation }
        '403': { description: Missing or invalid CSRF token }
        '429': { description: Too many requests }
  /season-notifications/unsubscribe:
    post:
      operationId: cancelSeasonNotification
      summary: Cancel a notification using a bearer token and a deliberate user action
      description: Requires X-XSRF-TOKEN. Repeated cancellation succeeds idempotently. An email already in flight may complete.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SeasonNotificationLink' }
      responses:
        '204': { description: Notification cancelled }
        '400': { description: Invalid token }
        '403': { description: Missing or invalid CSRF token }
        '429': { description: Too many requests }
  /moderation/season-notifications:
    get:
      operationId: getSeasonNotificationConfiguration
      summary: Moderator-only launch and configuration flags
      responses:
        '200':
          description: Non-secret configuration state
          content:
            application/json:
              schema:
                type: object
                required: [configured, seasonLive]
                properties:
                  configured: { type: boolean }
                  seasonLive: { type: boolean }
        '401': { description: Missing authenticated session }
        '403': { description: Moderator role required }
  /moderation/season-notifications/launch:
    post:
      operationId: queueSeasonLaunchNotifications
      summary: Moderator-only manual launch campaign
      description: Requires X-XSRF-TOKEN, a moderator session and the server season-live flag. Operators must verify ranked play is available first. Repeating the request queues no duplicates. Counts queued messages, not delivered inboxes.
      responses:
        '200':
          description: Campaign queued durably
          content:
            application/json:
              schema:
                type: object
                required: [queued]
                properties:
                  queued: { type: integer, minimum: 0 }
        '401': { description: Missing authenticated session }
        '403': { description: Moderator role or valid CSRF token required }
        '409': { description: Season-live flag is disabled }
        '503': { description: Email delivery has not been configured }
  /characters/me:
    get:
      operationId: getCurrentUserCharacters
      summary: Get the authenticated user's WoW Classic characters of level 60 or above in EU
      description: Requires a Battle.net session with wow.profile permission. Queries profile-classic1x-eu and profile-classic-eu using the user's token, merges WoW accounts, and returns characters of level 60 or above with their source namespace. A missing profile (404) in one branch does not hide the other branch. Never returns tokens or persists character associations.
      responses:
        '200':
          description: WoW accounts and their owned characters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WowAccountProfile'
        '401':
          description: Missing session or expired Battle.net token; sign in again
        '403':
          description: Missing wow.profile permission or Blizzard profile privacy restriction
        '404':
          description: Neither WoW Classic branch has an accessible profile in EU
        '502':
          description: Blizzard unavailable or returned an invalid response
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the current rbgs.io account and establish a CSRF cookie
      description: Returns no content when the browser has no authenticated session. Never returns provider tokens.
      responses:
        '200':
          description: Authenticated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '204':
          description: No authenticated session
  /auth/logout:
    post:
      operationId: logout
      summary: Invalidate the browser session
      description: Requires the X-XSRF-TOKEN header copied from the XSRF-TOKEN cookie.
      responses:
        '204':
          description: Session invalidated
        '403':
          description: Missing or invalid CSRF token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /health:
    get:
      operationId: getHealth
      summary: Check API availability
      responses:
        '200':
          description: API process is serving requests
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
components:
  schemas:
    SeasonNotificationLink:
      type: object
      required: [token]
      properties:
        token: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
    WowAccountProfile:
      type: object
      required: [wow_accounts]
      properties:
        wow_accounts:
          type: array
          items:
            type: object
            required: [id, characters]
            properties:
              id:
                type: integer
                format: int64
              characters:
                type: array
                items:
                  $ref: '#/components/schemas/WowCharacter'
    WowCharacter:
      type: object
      required: [namespace]
      properties:
        gender:
          anyOf:
            - $ref: '#/components/schemas/WowNamedReference'
            - type: 'null'
          description: Character gender from Blizzard (type MALE or FEMALE); null when unavailable.
        guild:
          anyOf:
            - $ref: '#/components/schemas/WowNamedReference'
            - type: 'null'
          description: Optional guild from the character profile; null when unguilded or unavailable.
        namespace:
          type: string
          enum: [profile-classic1x-eu, profile-classic-eu]
          description: Source Classic branch; part of the character's identity alongside account, realm and character IDs.
        avatarUrl:
          type: [string, 'null']
          format: uri
          description: Optional Blizzard portrait; null when character media is unavailable.
        insetUrl:
          type: [string, 'null']
          format: uri
          description: Optional wider Blizzard portrait; null when not provided.
        id:
          type: integer
          format: int64
        name:
          type: string
        level:
          type: integer
          minimum: 60
        realm:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
            slug:
              type: string
        playable_class:
          $ref: '#/components/schemas/WowNamedReference'
        playable_race:
          $ref: '#/components/schemas/WowNamedReference'
        faction:
          $ref: '#/components/schemas/WowNamedReference'
    WowNamedReference:
      type: object
      properties:
        id:
          type: [integer, 'null']
          format: int64
        name:
          type: [string, 'null']
        type:
          type: [string, 'null']
    Profile:
      type: object
      required: [id, displayName, region, role]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        region:
          type: string
          const: EU
        role:
          type: string
          enum: [USER, MODERATOR]
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string